+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Switcher: Changed wrapper span tag to div
+- DropDown: Fixed closing dropdown list on click outside an element in interactionMode="click"
 
 ## [0.31.1] - 2020-09-28